requirements.txt
streamlit
pandas
altair
plotly.express
sqlalchemy